chore: optmise the code
