+++ src/RandomReveal.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+import { Props } from './types'
+import { useRandomReveal } from '.'
+
+export const RandomReveal = (props: Props) => {
+  const nextCharacters = useRandomReveal(props)
+
+  return <>{nextCharacters}</>
+}
