+++ src/easing.ts
