--- src/hooks/index.ts
@@ -1,2 +1,0 @@
-export { useRandomCharacters } from './useRandomCharacters'
-export { useRandomReveal } from './useRandomReveal'
