+++ src/index.ts
@@ -1,0 +1,2 @@
+export { RandomReveal } from './RandomReveal'
+export { useRandomReveal } from './useRandomReveal'
