--- src/index.tsx
@@ -1,4 +1,0 @@
-
-export { RandomReveal } from './components'
-export { useRandomReveal } from './hooks'
-
