--- src/utils/characters.ts
@@ -1,31 +1,0 @@
-import { Characters, CharactersRequired } from '../types'
-
-export const defaultIgnoreCharacterSet: Characters = []
-export const defaultCharacterSet: CharactersRequired = [
-	'a',
-	'b',
-	'c',
-	'd',
-	'e',
-	'f',
-	'g',
-	'h',
-	'i',
-	'j',
-	'k',
-	'l',
-	'm',
-	'n',
-	'o',
-	'p',
-	'q',
-	'r',
-	's',
-	't',
-	'u',
-	'v',
-	'w',
-	'x',
-	'y',
-	'z',
-]
