--- src/utils/getRandomCharacter.ts
@@ -1,5 +1,0 @@
-import { CharactersRequired, Character } from '../types'
-
-export const getRandomCharacter = (set: CharactersRequired): Character => {
-	return set[Math.floor(Math.random() * set.length)]
-}
