--- src/utils/index.ts
@@ -1,4 +1,0 @@
-export { defaultCharacterSet, defaultIgnoreCharacterSet } from './characters'
-export { easings } from './easing'
-export { getCharactersData } from './getCharactersData'
-export { getRandomCharacter } from './getRandomCharacter'
